# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Extra dice purchased with XP for an attribute
 */
export interface DicePool {
  d4: number;
  d6: number;
  d10: number;
  d20: number;
}

export interface AttributeData {
  value: number;
  dicePool: DicePool;
}

export interface XPLogEntry {
  id: string;
  attribute: string;
  diceType: string;
  cost: number;
  timestamp: string;
}

export interface CharacterInput {
  playerName: string;
  pirateName: string;
  origin: string;
  specialty: string;
  vigor: AttributeData;
  agility: AttributeData;
  cunning: AttributeData;
  charisma: AttributeData;
  spirit: AttributeData;
  maxHp: number;
  currentHp: number;
  berries: number;
  xpTotal: number;
  logbook: string;
  xpLog: XPLogEntry[];
}

export type Character = CharacterInput & {
  id: number;
  updatedAt: string;
};

export interface ShipItem {
  id: string;
  name: string;
  quantity: number;
}

export interface ShipItemInput {
  name: string;
  quantity: number;
}

export interface ShipInput {
  name: string;
  maxHull: number;
  currentHull: number;
  treasury: number;
  items: ShipItem[];
}

export type Ship = ShipInput & {
  code: string;
  updatedAt: string;
};