# lib/api-zod/src/generated/types/attributeData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */
import type { DicePool } from "./dicePool";

export interface AttributeData {
  value: number;
  dicePool: DicePool;
}