# lib/api-zod/src/generated/types/characterInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */
import type { AttributeData } from "./attributeData";
import type { XPLogEntry } from "./xPLogEntry";

export interface CharacterInput {
  playerName: string;
  pirateName: string;
  origin: string;
  specialty: string;
  vigor: AttributeData;
  agility: AttributeData;
  cunning: AttributeData;
  charisma: AttributeData;
  spirit: AttributeData;
  maxHp: number;
  currentHp: number;
  berries: number;
  xpTotal: number;
  logbook: string;
  xpLog: XPLogEntry[];
}