# lib/api-zod/src/generated/types/dicePool.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Extra dice purchased with XP for an attribute
 */
export interface DicePool {
  d4: number;
  d6: number;
  d10: number;
  d20: number;
}