# lib/api-zod/src/generated/types/ship.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */
import type { ShipInput } from "./shipInput";

export type Ship = ShipInput & {
  code: string;
  updatedAt: string;
};