# lib/api-zod/src/generated/types/shipInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */
import type { ShipItem } from "./shipItem";

export interface ShipInput {
  name: string;
  maxHull: number;
  currentHull: number;
  treasury: number;
  items: ShipItem[];
}