# lib/api-zod/src/generated/types/shipItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface ShipItem {
  id: string;
  name: string;
  quantity: number;
}