# lib/api-zod/src/generated/types/shipItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface ShipItemInput {
  name: string;
  quantity: number;
}