# lib/api-zod/src/generated/types/xPLogEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * One Piece RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface XPLogEntry {
  id: string;
  attribute: string;
  diceType: string;
  cost: number;
  timestamp: string;
}